build/summary: print all procedure info into output.json
We now print all traces into output.json, too.
